Add pointer event injection with win32u
It was (mostly) obvious how the syscall interface of pointer injection
works based on the public API. This research (hopefully) will allow us
to implement these syscalls in Wine.

I'm not sure if implementations of these syscalls will be allowed, but
it seems like existing user32 code does not call directly into
wineserver, yet for input injection and WM_POINTER in general, we
probably do need to add new wineserver calls. So it seems most correct
to implement the syscall interface, which then calls wineserver.